stress-cpu: ensure i64 is stashed in stress_cpu_intconversion
i64 was being updated and not used leading it to be optimized out.
Ensure the value in i64 is stashed away to force the computation.

+++ stress-cpu.c
@@ -1779,7 +1779,7 @@ static void stress_cpu_intconversion(const char *name)
 		i32 = -(int64_t)(uint32_t)-(int16_t)(uint64_t)i16_sum;
 		i32_sum += i32;
 		i64 = -(int64_t)(uint32_t)-(int16_t)(uint64_t)i32_sum;
-		i64_sum += i32;
+		i64_sum += i64;
 	}
 	stress_uint16_put(i16_sum);
 	stress_uint32_put(i32_sum);
